Move example sketches to appropriately named folder
The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.

I renamed examples/bitBrick_bitHand.ino to examples/bitBrick_bitHand2.ino because there was an existing sketch of that name with different code.